change Requires.jl code to package extension
