+++ Project.toml
@@ -12,7 +12,17 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 
 [compat]
 Latexify = "0.15, 0.16"
+Plots = "1"
 Preferences = "1.3"
 RecipesBase = "1"
 Requires = "1"
 julia = "1.6"
+
+[extensions]
+PlotsExt = "Plots"
+
+[extras]
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+
+[weakdeps]
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
